# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API para confeitaria premium Docinho & Cia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sweets:
    get:
      operationId: getSweets
      tags: [sweets]
      summary: List all docinhos
      responses:
        "200":
          description: List of sweets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sweet"
    post:
      operationId: createSweet
      tags: [sweets]
      summary: Add a new sweet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SweetInput"
      responses:
        "201":
          description: Created sweet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sweet"
  /sweets/{id}:
    put:
      operationId: updateSweet
      tags: [sweets]
      summary: Update a sweet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SweetUpdate"
      responses:
        "200":
          description: Updated sweet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sweet"
    delete:
      operationId: deleteSweet
      tags: [sweets]
      summary: Delete a sweet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /testimonials:
    get:
      operationId: getTestimonials
      tags: [testimonials]
      summary: List all testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Add a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    put:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Sweet:
      type: object
      required: [id, name, description, imageUrl, sortOrder, visible]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    SweetInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    SweetUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
...
    Testimonial:
      type: object
      required: [id, name, text, sortOrder, visible]
      properties:
        id:
          type: integer
        name:
          type: string
        text:
          type: string
        photoUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    TestimonialInput:
      type: object
      required: [name, text]
      properties:
        name:
          type: string
        text:
          type: string
        photoUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    TestimonialUpdate:
      type: object
      properties:
        name:
          type: string
        text:
          type: string
        photoUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
...